remove fix

+++ commands/remove.js
@@ -3,14 +3,21 @@ import db from '../services/db';
 import colors from '../models/colors';
 
 export default async function (interaction) {
+	if (interaction.guildId !== '669867414409969664') return;
+
 	const id = interaction.options.getString('id');
+
+	if (!id) return interaction.reply({ content: 'ID не валидный!' });
+
 	const guild = await bot.guilds.fetch(id);
+
+	if (!guild) return interaction.reply({ content: 'Гильдия не найдена!' });
 
 	await guild.leave();
 	await db.query('DELETE FROM server WHERE id = ?', [id]);
 
 	const embed = new MessageEmbed()
-		.setAuthor(`Успешно! Сервер ${guild.name} удален!`, guild.iconURL())
+		.setAuthor(`Успешно! Сервер «${guild.name}» удален!`, guild.iconURL())
 		.setColor(colors.green)
 		.setTimestamp()
 		.setFooter(interaction.user.tag);
